refactor(#4): extract Address interface
reason: using the type from the API response directly creates a direct (compile-time) dependency of frontend components on the API result.
Changes to the API would not only affect the page (where we fetch from the API) but all other components.
This change makes child components independent of the API, while still having type safety for API clients.
