+++ src/components/common/address/Address.ts
@@ -1,0 +1,13 @@
+export type Address = {
+  location: {
+    x: number;
+    y: number;
+  };
+  attributes: {
+    Type: string;
+    Place_addr: string;
+    PlaceName: string;
+    Country: string;
+    CntryName: string;
+  };
+};
